regression_1000: Fix 1020 test when Pseudo TA is disabled
It was found that OP-TEE can return BAD_PARAMETERS error instead of
NOT_FOUND when pseudo TA is disabled in OP-TEE. In this case
EXPECT_TEEC_SUCCESS result should be checked and error returned if we
unable to open session. Not checking this result leads to an Exception
of invoking command on bad session_id.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>
Acked-by: Sergiy Kibrik <sergiy_kibrik@epam.com>