src: regression_6000: fix memory corruption on test 6018
Allocate big buffer for test 6018 in a global section because there
is a risk that if there is not enough stack size - it may overwrite
globals needed for ZTEST. The corruption of failed_expectation and
test_result globals was observed.
With test_result the corruction isn't significant because it is
overwritten right after test finish. But failed_expectation corruption
causing weird behaviour when Test was passed but it is marked as failed.

Signed-off-by: Oleksii Moisieiev <oleksii_moisieiev@epam.com>